refactor: 패키지 구조 변경, common 패키지 생성
- 공통적으로 사용되는 것들을 묶어 common 패키지로 이동했습니다.
