+++ src/main/java/wooteco/auth/api/AuthControllerAdvice.java
@@ -7,7 +7,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import wooteco.auth.exception.AuthenticationException;
-import wooteco.dto.ErrorResponse;
+import wooteco.common.dto.ErrorResponse;
 
 @Order(Ordered.HIGHEST_PRECEDENCE)
 @RestControllerAdvice(assignableTypes = {AuthController.class, MemberController.class})
