+++ src/main/java/wooteco/common/config/WebConfig.java
@@ -1,4 +1,4 @@
-package wooteco.config;
+package wooteco.common.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
