+++ src/main/java/wooteco/subway/api/SubwayControllerAdvice.java
@@ -6,7 +6,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import wooteco.dto.ErrorResponse;
+import wooteco.common.dto.ErrorResponse;
 
 import java.sql.SQLException;
